+++ release_notes.md
@@ -4,3 +4,8 @@
 
 ## New Features
 
+- units support: when loading a VBRc .mat file, arrays will be loaded as unyt arrays if possible (requires VBRc version >= 1.0.0)
+
+## Bug Fixes
+
+None